requirements.txt
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.1.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
django>=4.2.0
tiktoken>=0.9.0